modified file

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -8,7 +8,7 @@
 */
 int main(void)
 {
-	unsigned long int n1, n2, sum, i;
+	unsigned long long int n1, n2, sum, i;
 
 	n1 = 0;
 	n2 = 1;
